+++ tcs/Heliostat_AzElAod.cpp
@@ -130,7 +130,7 @@ public:
 	Heliostat3DInterp( tcscontext *cst, tcstypeinfo *ti)
 		: tcstypeinterface( cst, ti)
 	{
-		n_hel = std::numeric_limits<double>::quiet_NaN();
+		n_hel = (int)std::numeric_limits<double>::quiet_NaN();
 		q_start = std::numeric_limits<double>::quiet_NaN();
 		p_run = std::numeric_limits<double>::quiet_NaN();
 		v_wind_max = std::numeric_limits<double>::quiet_NaN();
@@ -147,7 +147,7 @@ public:
 	virtual int init()
 	{
 		// Read in parameters
-		n_hel = value( P_n_hel );					// [-] Number of heliostats
+		n_hel = (int)value( P_n_hel );					// [-] Number of heliostats
 		q_start = value( P_q_start ) * 3600.0;		// [kJ] convert from kWe-hr
 		p_run = value( P_p_run ) * 3600.0;			// [kJ/hr] convert from kWe
 		v_wind_max = value( P_v_wind_max );			// [m/s] Wind speed at which heliostats are stowed
